requirements.txt
# Kubera Portfolio Agent Orchestration - Fixed for Python 3.13
# Let pip resolve compatible versions

# === AI/LLM Frameworks ===
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# CrewAI - using latest compatible version
crewai>=1.0.0
crewai-tools>=1.0.0

# === OpenAI/OpenRouter ===
openai>=1.50.0

# === Data & Market Data ===
yfinance>=0.2.48
pandas>=2.2.0
numpy>=2.0.0
# Note: pandas-ta not compatible with Python 3.13, using ta-lib alternative
ta>=0.11.0

# === Web Scraping & APIs ===
requests>=2.32.0
beautifulsoup4>=4.12.0
alpha-vantage>=2.3.1

# === Telegram Bot ===
python-telegram-bot>=21.0

# === Utilities ===
python-dotenv>=1.0.0
tenacity>=9.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# === Testing (Optional) ===
pytest>=8.0.0
pytest-asyncio>=0.24.0
